+++ webpack.mix.js
@@ -1,4 +1,5 @@
 const mix = require('laravel-mix');
 
 mix.js('resources/js/app.js', 'public/js')
-    .sass('resources/sass/app.scss', 'public/css');
+    .sass('resources/sass/app.scss', 'public/css')
+    .version();